+++ src/assets/ba_data/python/babase/_hooks.py
@@ -33,16 +33,28 @@ def reset_to_main_menu() -> None:
         logging.warning('reset_to_main_menu: no-op due to classic not present.')
 
 
-def set_config_fullscreen_on() -> None:
+def store_config_fullscreen_on() -> None:
     """The OS has changed our fullscreen state and we should take note."""
     _babase.app.config['Fullscreen'] = True
     _babase.app.config.commit()
 
 
-def set_config_fullscreen_off() -> None:
+def store_config_fullscreen_off() -> None:
     """The OS has changed our fullscreen state and we should take note."""
     _babase.app.config['Fullscreen'] = False
     _babase.app.config.commit()
+
+
+def set_config_fullscreen_on() -> None:
+    """Set and store fullscreen state"""
+    _babase.app.config['Fullscreen'] = True
+    _babase.app.config.apply_and_commit()
+
+
+def set_config_fullscreen_off() -> None:
+    """The OS has changed our fullscreen state and we should take note."""
+    _babase.app.config['Fullscreen'] = False
+    _babase.app.config.apply_and_commit()
 
 
 def not_signed_in_screen_message() -> None: