+++ src/ballistica/base/python/base_python.h
@@ -38,6 +38,8 @@ class BasePython {
     kConfig,
     kAppOnNativeBootstrappingCompleteCall,
     kResetToMainMenuCall,
+    kStoreConfigFullscreenOnCall,
+    kStoreConfigFullscreenOffCall,
     kSetConfigFullscreenOnCall,
     kSetConfigFullscreenOffCall,
     kNotSignedInScreenMessageCall,